% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{changeSign_loads_spca}
\alias{changeSign_loads_spca}
\title{changes the signs to loadings in spca object}
\usage{
changeSign_loads_spca(spca_obj, index_to_change)
}
\arguments{
\item{spca_obj}{an object of class spca}

\item{index_to_change}{vector of indices of loaidngs to change sign}
}
\value{
the spca object with changed signs
}
\description{
laodings, contributions, correlation between components and scores are changed
}
