% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{crime_data}
\alias{crime_data}
\alias{crime_var_lookup}
\title{Crime data and variable lookup}
\format{
\describe{
  \item{crime_data}{A numeric data frame with 1,994 observations and 99 variables scaled to mean zero and unit variance.}
  \item{crime_var_lookup_df}{A data frame giving variable definitions for \code{crime_data}.}
}
}
\source{
\url{http://archive.ics.uci.edu/ml/machine-learning-databases/communities/}
}
\description{
\code{crime_data} combines socio-economic data from the 1990 US Census, law enforcement
data from the 1990 US LEMAS survey, and crime data from the 1995 FBI UCR.
}
\details{
In this package, the variable order was rearranged and 25 variables with missing values were removed.
The variables in \code{crime_data} are scaled to mean zero and unit variance.

\code{crime_var_lookup_df} provides variable definitions corresponding to the columns of
\code{crime_data}.
}
\references{
Redmond, M. A. and Baveja, A. (2002). A Data-Driven Software Tool for Enabling Cooperative
Information Sharing Among Police Departments. \emph{European Journal of Operational Research},
141, 660--678.
}
\keyword{dataset}
