% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods_For_Class.R
\name{is.spca}
\alias{is.spca}
\title{\strong{Verifies if an object is of class spca}}
\usage{
is.spca(x)
}
\arguments{
\item{x}{Any object suspected of being of class spca.}
}
\value{
Logical: TRUE if object is of class spca and contains the minimal configuration, 
FALSE otherwise.
}
\description{
\strong{Verifies if an object is of class spca}
}
\details{
The function carries out checks for the minimal components of an
\emph{spca} object. It checks the presence and mode of
 a matrix of loadings named \emph{loadings} 
 a vector of variance explained by the components named \emph{vexp} 
 a vector of variance explained by the PCss named \emph{vexpPC}
an integer declaring the number of components named \emph{ncomps}.
If any of these is missing the object is deemed as not \emph{spca} and a
warning is issued, even if the object's class is \emph{spca}.
}
