% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsspca.R
\name{lsspca}
\alias{lsspca}
\title{Computes LS SPCA using different selection methods}
\usage{
lsspca(
  X,
  alpha = 0.95,
  ncomps = 0,
  ncbyvexp = 1,
  method = c("p", "c", "u"),
  subsearch = c("stepwise", "backward", "forward"),
  maxcard = 0,
  force_in = NULL,
  force_out = NULL,
  scalex = FALSE,
  mkvif = FALSE
)
}
\arguments{
\item{X}{data frame or matrix with data.}

\item{alpha}{= 0.95. Real in [0,1] (optional if ncbyvexp < 1). Percentage of variance of the PCs explained by the sparse component.}

\item{ncomps}{= 0 Number of components to compute (if passed 0, reset to the number of variables)}

\item{ncbyvexp}{= 1 (optional if ncomps < 1) real in [0, 1] compute components until this fraction of total vexp explained. whichever satisfied first ncomps.}

\item{method}{how lsspca is computed}

\item{maxcard}{integer (optional) a vector or one integer. Missing values filled with last value.}

\item{force_in}{NULL or list or vector of indices must be in component. not for lasso. [NULL]}

\item{force_out}{NULL or list or vector of indices cannot be in component. [NULL]}

\item{scalex}{= FALSE if TRUE the variables are scaled to unit length. Variables are automatically scaled to zero mean (with warning) if they are not.}

\item{mkvif}{= FALSE, if true computes the variance inflation factors for each variable selected by each component}

\item{subset}{search one of "stepwise", "backward" or "forward". Default "stepwise".}
}
\value{
an spca object
}
\description{
For each component, the variables are selected so as to explain
a percentage \emph{alpha} of the vexp by the corresponding principal component.
}
