% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{make_vif.spca}
\alias{make_vif.spca}
\title{makes VIF for an spca object}
\usage{
make_vif.spca(spca_obj, M, intercept = FALSE)
}
\arguments{
\item{spca_obj}{surprisingly, an spca object}

\item{M}{data matrix or variance/correlation matrix (same result)}

\item{intercept}{if include intercept, use TRUE only for data matrix not centered.}
}
\value{
a list of vectors with vif values for each component
}
\description{
experimental use with care, uses only vifSC because vifXC  slow
}
