% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_support.R
\name{make_vif_R}
\alias{make_vif_R}
\title{\strong{Computes variance inflaction factors for variables in spca components.}}
\usage{
make_vif_R(
  spca_obj,
  M,
  intercept = FALSE,
  pseudo = FALSE,
  prn = TRUE,
  digits = 0
)
}
\arguments{
\item{spca_obj}{An spca object, typically from SPCA.}

\item{M}{the data or covariance/correlation matrix}

\item{intercept}{Logical: should the intercept be included in the model?}

\item{pseudo, }{should compute vifs using generalised inverse?
Only when M is data matrix.}

\item{prn}{Logical. Should the function print the results?}

\item{digits.}{Number of decimal figures to print and return. See details.}
}
\value{
A list of ordered vifs for each component.
}
\description{
result provided by SPCA
}
\details{
A wrapper for vifSC in C++. May run into problems if the variables
are perfectly multicollinear, in that c(ase use \code{pseudo = TRUE}.  
Vifs values are named with \code{colnames(M)}. 
If \code{digits} is 0, loadings are returned unrounded and printed (if enabled) 
to two digits. If \code{digits} > 0, loadings are returned rounded to that value.
}
