% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods_For_Class.R
\name{new_spca}
\alias{new_spca}
\title{\strong{Constructs an object of class spca from a set of loadings}}
\usage{
new_spca(A, S, X = NULL, method = NULL)
}
\arguments{
\item{A}{real matrix, a matrix of loadings.}

\item{S}{the variance or correlation matrix from which the loadings where computed.}

\item{X}{real matrix, the data matrix from which the loadings where computed, optional}

\item{method}{string, the name of the method used to compute the loadings, optional.}
}
\value{
An object is of class spca with several added components, 
FALSE otherwise.
}
\description{
\strong{Constructs an object of class spca from a set of loadings}
}
\details{
The \code{spca} object can be used as argument to the spca methods.
 It contains \describe{ 
 \item{loadings}{the matrix of loadings.}
 \item{contributions}{the matrix of loadings scaled to unit L1 norm.} 
 \item{ncomps}{the number of components.}
 \item{cardinality}{the vector with the number of nonzero loadings in 
 each each component.}
 \item{ind}{a list with the indices of the variables corresponding the nonzero loadings.}
 \item{vexp}{a vector of percent variance explained by the components.} 
 \item{cvexp}{a vector of percent cumulative variance explained by the components.}
 \item{vexpPC}{a vector of variance explained by the PCs.}
 \item{rvexp}{a vector of proportion of cumulative variance explained over that explained by the 
 corresponding PCs.}
 \item{loadlist}{a list with only the nonzero loadings.}
 \item{scores}{the components' scores, if X is passed.} 
 \item{corComp}{A matrix of the correlations between components.}
 \item{vif}{a list of variance inflation factors for the variables in each components, 
 if X is passed.}
 \item{method}{the string passed as such, if method is passed.}
}
}
\seealso{
\code{\link{PMlsspca-package}} for a description of an \code{spca} object.
}
