% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pca}
\alias{pca}
\title{\strong{Computes principal components solutions}}
\usage{
pca(
  M,
  ncomps,
  centerdata = FALSE,
  scaledata = FALSE,
  only.values = FALSE,
  screeplot = FALSE,
  kaiser_line = F,
  kaiser.print = FALSE
)
}
\arguments{
\item{M}{A data matrix or correlation or covariance matrix.}

\item{ncomps}{Integer: number of loadings to retain. If missing all loadings are retained.}

\item{centerdata}{= FALSE, should variables be centered to zero mean automatically done if any mean is not zero}

\item{scaledata}{= FALSE , should variables be centered to unit variance?}

\item{only.values}{Logical: should only the eigenvalues be computed?}

\item{screeplot}{Logical: should the screeplot be plotted?}

\item{kaiser.print}{Logical: should the kaiser rule be computed, printed and
returned?.}

\item{kaiser.line}{Logical: adds a horizontal line to the screeplot at y = 1}
}
\value{
An object of class \emph{spca} is returned, which contains:
\describe{
\item{loadings}{Matrix with the loadings scaled to unit \eqn{L_2} norm in 
the columns.} 
\item{contributions}{Matrix of loadings scaled to unit \eqn{L_1} norm.} 
\item{ncomps}{integer number of components computed.} 
\item{cardinality}{Vector with the cardinalities of each loadings.} 
\item{ind}{List with the indices of the non-zero loadings for each component.} 
\item{vexp}{Vector with the \% variance explained by each component.} 
\item{cvexp}{Vector with the \% cumulative variance explained by each component.} 
\item{vexpPC}{Vector with the \% variance explained by each principal component.} 
\item{rcvexp}{Vector with the proportion of cumulative variance explained by each 
component over the cumulative variance explained by the corresponding PCs.}
\item{Call}{The called with its arguments.}
\item{kaiser}{The number of eigenvalues larger than one, if \code{kaiser.print = TRUE}.}
}
}
\description{
Creates PCA output as an \link{PMlsspca-package} object
which can run on the spca utilities. .
}
\details{
\emph{ncomps} is just the number of components retained from the full eigen
decomposition, doesn't speed up the function. \emph{only.values} does not compute the loadings and is more efficient. Kaiser rule determines the number of components as the number of eigenvalues larger than one. It should
be used only for correlation matrices, if called on a covariance matrix a
warning is generated.
}
\seealso{
See also \code{\link{print.spca}, \link{summary.spca}}
######=============================================================  
######=============================================================
}
