% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods_For_Class.R
\name{plot.spca}
\alias{plot.spca}
\title{\strong{Plots loadings from an spca object}}
\usage{
\method{plot}{spca}(
  spca_obj,
  nplot,
  plotcontributions = TRUE,
  onlynonzero = TRUE,
  varnames = TRUE,
  circular = TRUE,
  pcloadings = NULL,
  colourscale = c("cbb", "printsafe", "bw", "ggplot"),
  returnplot = FALSE,
  produceplot = TRUE
)
}
\arguments{
\item{spca_obj}{An spca object.}

\item{nplot}{The number of components to be plotted. Default all. If an
itenger is passed, it is set to 1:nplot.}

\item{plotcontributions}{Logical: should the loading be scaled as percentages?}

\item{onlynonzero}{Logical: should only the non-zero loadings be plotted?}

\item{varnames}{Hybrid: if TRUE the variable names from the laoding matrix are added to the plot. 
If a vector with the names is passed, those names are used as labels, if FALSE no labels
are added.}

\item{circular}{If TRUE a circular barplot is produced. Otherwise a linear one.}

\item{pcloadings}{An spca object containing the PCA loadings, typically obtained
with the function pca.}

\item{returnplot}{Logical, if the plot must be returned}

\item{produceplot}{Logical, if the plot is to be plotted (useful when only 
the plot needs to be retuned)}

\item{colorscale}{can take values c("cbb", "printsafe", "bw", "ggplot")
or a pallettte with colour codes. See details for explanation.}
}
\value{
If returnplot = TRUE the ggplot2 object
}
\description{
Plots loadings from an spca solutions.
}
\details{
The loadings are plotted as circular or linear barplots. For large
matrices it is reccommended to set onlynonzero = TRUE and variablesnames = FALSE. 
The sparse loadings can be plotted together with the PC's for comparison.

colourscaler takes \emph{cbb} for colorblind (with black) friendly, 
\emph{printsafe} for colourblind and printer friendly, \emph{bw} for gray tones and 
\emph{ggplot} for the default ggplot2 scale. A vector with colour hexadecimal RGB 
codes can be passed.
}
\examples{

\dontrun{
library(SPCA3)
data("hitters_cor")

## Four uncorrelated spca components each explaining 95\% of PC's vexp
hit_spca = SPCA(M = hitters_cor, alpha = 0.95, ncomps = 4)

## plot circular barplots of all SPCA contributions 
## with colourblind friendly colours
plot(hit_spca, nplot = 4, 
     onlynonzero = FALSE, 
     colourscale = "cbb"
    )

## plot circular barplots of nonzero SPCA contributions with printer safe colours
plot(hit_spca, nplot = 4, 
     colourscale = "printsafe"
    )

## plot linear barplots of nonzero SPCA contributions with custom colours
Set1 = c("#E41A1C", "#377EB8", "#4DAF4A", "#984EA3", "#FF7F00", "#FFFF33",
         "#A65628", "#F781BF", "#999999")#, brewer
plot(hit_spca, nplot = 4, 
     circular = FALSE, 
     colourscale = Set1
    )

## plot SPCA and PCA contributions for 3 components and ggplot2 colours
hit_ee = eigen(hit_c) ## loadings PCA
plot(hit_spca, nplot = 3, 
     colourscale = "ggplot",
     pcloadings = hit_ee$vectors
     ) 
}
}
\references{
The circular barplots I discovered on the \emph{r-graph-gallery} 
website \url{https://www.r-graph-gallery.com/all-graphs/}. The cbb colur pallette I found on
the \emph{cookbook-r} website, \url{http://www.cookbook-r.com/Graphs/Colors_(ggplot2)/}. The 
printsafe pallette is the OrRd from the colorbrewer2 website \url{http://colorbrewer2.org/}
}
\seealso{
Examples in \code{\link{spcabe}} and \code{\link{spcabb}}. For plotting two or
more spca solutions together see \code{\link{compare}}.
}
