% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods_For_Class.R
\name{plot_vexp.spca}
\alias{plot_vexp.spca}
\title{\strong{Plots vexp from an spca object}}
\usage{
plot_vexp.spca(
  spca_obj,
  nplot,
  vexpPC = FALSE,
  rcvexp = FALSE,
  alpha = NULL,
  labelcardinality = TRUE,
  colourscale = c("cbb", "bw", "ggplot"),
  returnplot = FALSE,
  produceplot = TRUE
)
}
\arguments{
\item{spca_obj}{An spca object.}

\item{nplot}{Integer The number of cvexps to be plotted. Default all.}

\item{vexpPC}{Logical: should the cvexp by the pcs be also plotted?
if == TRUE the object must contain a \emph{vexpPC} vector}

\item{rcvexp}{Logical: plot the vexps relative to the vexp by the PC?
Cannot be TRUE when vexpPC is also TRUE}

\item{alpha}{the $R^2$ value used for stopping variable selection. If
 passed with rcvexp == TRUE, this will be plotted as alower bound. 
If set to \code{NULL} (default) it won't be plotted}

\item{labelcardinality}{Logical If TRUE the cardinality of each component
are added to the plot}

\item{returnplot}{Logical, if the plot must be returned}

\item{produceplot}{Logical, if the plot is to be plotted (useful when only 
the plot needs to be retuned)}

\item{clolourscale}{character. Colors for plotting, Only for vexpPC = TRUE.
If "cbb" colour blind friendly combination, if "bw" gray shades, anything else
ggplot2 default colour scheme.}
}
\value{
If returnplot = TRUE the ggplot2 object
}
\description{
Plots the cumulative variance explained by spca components.
}
\details{
The cvexp can be plotted as percentage of the total variance or relative to the cvexp
by the PCs. It is also possible to plot the cvexp by the sparse  components and PCs together.
}
\examples{

\dontrun{
library(SPCA3)
data("hitters_cor")

## Four uncorrelated spca components each explaining 95\% of PC's vexp
hit_spca = SPCA(M = hitters_cor, alpha = 0.95, ncomps = 4)

# plot the \% variance eplained with cardinality labels
plot_vexp.spca(hit_spca, nplot = 4)

# plot the variance eplained relative to vexp PCs without cardinality labels
## add a line for minimum rcvexp
plot_vexp.spca(hit_spca, nplot = 4, rcvexp = TRUE, alpha = 0.95, 
         labelcardinality = FALSE)

# plot the \% variance eplained with that of PCs with cardinality labels
# in gray shades
plot_vexp.spca(spca_obj = hit_spca, vexpPC = TRUE,
               labelcardinality = TRUE, colourscale = "bw")

# plot the \% variance eplained with that of PCs with cardinality labels
# only 3 components with custom colours
plv = plot_vexp.spca(spca_obj = hit_spca, nplot = 3, vexpPC = TRUE,
               labelcardinality = TRUE, returnplot = TRUE, 
               produceplot = FALSE, colourscale = "none")
plv + scale_colour_manual(values = c("red", "green"))
}
}
\seealso{
Examples in \code{\link{SPCA}}.
}
