\name{PMlsspca-package}
\alias{PMlsspca-package}
\alias{PMlsspca}
\docType{package}
\title{
\packageTitle{PMlsspca}
}
\description{
\packageDescription{PMlsspca}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{PMlsspca}
\packageIndices{PMlsspca}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{PMlsspca}

Maintainer: \packageMaintainer{PMlsspca}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
