% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods_For_Class.R
\name{print.spca}
\alias{print.spca}
\title{\strong{Prints the sparse loadings from an spca object}}
\usage{
\method{print}{spca}(
  spca_obj,
  cols,
  only.nonzero = TRUE,
  perc = TRUE,
  digits = 3,
  thresh = 0.001,
  rtn = FALSE,
  namescomp = NULL,
  ...
)
}
\arguments{
\item{spca_obj}{An spca object.}

\item{cols}{A vector indicating which components should be printed. Default
all. If an iteger is passed, it is set to 1:cols.}

\item{only.nonzero}{Logical: if = TRUE only the nonzero loadings are printed.
otherwise all loadings are printed.}

\item{perc}{Logical: should the loadings be standardised to unit \eqn{L_1}
norm (and printed as percentage contributions)?}

\item{digits}{Integer: number of decimal figures.}

\item{thresh}{Value below which loadings are considered zero and not
printed.}

\item{rtn}{Logical: should the formatted (text) table be returned?}

\item{namescomp}{A vector of names for the components. If NULL assigned as
"Comp j"}

\item{...}{Additonal arguments for generic print, additional arguments will
generate an error.}
}
\value{
If rtn = TRUE, it returns a text table formatted as specified by the
arguments.
}
\description{
Prints sparse loadings omitting the zero ones and giving the cumulative
variance explained.
}
\note{
This is a wrapper for the main function in which the "dots" are disabled
so that only exact (or partial) prescribed arguments can be entered.
}
\examples{

\dontrun{
### example print.spca ========
library(SPCA3)
data("hitters_cor")

## Four uncorrelated spca components each explaining 95\% of PC's vexp
hit_spca = SPCA(M = hitters_cor, alpha = 0.95, ncomps = 4)

## default print. Prints the nonzero \% contributions with 1 decimal
print(hit_spca)

## Prints the all loadings with 2 decimals
print(hit_spca, only.nonzero = FALSE, digits = 2, perc = F)

## Prints the nonzero \% contributions without decimals 
## (since \% values it means 2 digits)
print(hit_spca,  digits = 2)

## prints contributions larger than 25\%
print(hit_spca,  thresh = 0.25)
}
}
\seealso{
Examples in \link{spcabb} and \link{spcabe}.
}
