% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{replicate_article_examples}
\alias{replicate_article_examples}
\title{Replication script for the article}
\usage{
replicate_article_examples()
}
\value{
A character string giving the path to `Replicate_Examples.R` in the installed package (inst/scripts/"Replicate_Examples.R).
}
\description{
Returns the path to the script that reproduces the examples reported in the article.
}
