% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{roundl}
\alias{roundl}
\title{Rounds  a list}
\usage{
roundl(li, d = 2)
}
\arguments{
\item{li}{a list of numerical objects}

\item{d}{nuber of digits}
}
\value{
rounded list
}
\description{
prints or return a rounded list
}
