% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsspca.R
\name{lsspca}
\alias{lsspca}
\title{\strong{Computes LS SPCA using different selection methods}}
\usage{
lsspca(
  X,
  alpha = 0.95,
  ncomps = 0,
  ncbyvexp = 1,
  method = "p",
  varselection = c("stepwise", "backward", "forward"),
  maxcard = 0,
  force_in = NULL,
  force_out = NULL,
  scalex = FALSE,
  mkvif = FALSE
)
}
\arguments{
\item{X}{dataframe or matrix with data.}

\item{alpha}{= 0.95. Real in [0,1] (optional if ncbyvexp < 1). Minimum \code{R^2} of the regression of the residual PC for variable selection.}

\item{ncomps}{= 0 Number of components to compute (if passed 0, reset to the number of variables)}

\item{ncbyvexp}{= 1 (optional if ncomps < 1) real in [0, 1] compute components until this fraction of total vexp explained. whichever satisfied first ncomps.}

\item{method}{a character vector, one of "p", "c", "u", specifying how the lsspca loadings are computed for each component. The last value will be applied to all remaining components.}

\item{varselection}{one of "stepwise", "backward" or "forward". Default "stepwise". Initial letter enough.}

\item{maxcard}{integer (optional) a vector or one integer. Missing values filled with last value.}

\item{force_in}{NULL or list or vector of indices must be in component.}

\item{force_out}{NULL or list or vector of indices cannot be in component.}

\item{scalex}{= FALSE if TRUE the variables are scaled to unit length. Variables are automatically scaled to zero mean (with warning) if they are not.}

\item{mkvif}{= FALSE, if true computes the variance inflation for each variable selected  in every component May slowdown.}
}
\value{
an \code{\link{spca_object}}.
}
\description{
For each component, the variables are selected so as to explain
a percentage \emph{alpha} of the vexp by the corresponding principal component.
}
\details{
\code{alpha} controls the $R^2$. The function may be slow because it uses the package leaps for variable selection. This is to keep results the same as in the paper.
}
