% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMlsspca-package.R
\docType{package}
\name{PMlsspca-package}
\alias{PMlsspca}
\alias{PMlsspca-package}
\title{PMlsspca: Least Squares Sparse PCA utilities}
\description{
Tools to compute LSSPCA solutions and reproduce the results reported in the accompanying article.
}
\details{
The package provides functions to compute Least Squares Sparse Principal Components Analysis (LSSPCA) solutions, where sparsity is imposed while targeting PCA’s least-squares reconstruction objective. This release accompanies the related article and is intended  to support full reproduction of the results reported therein.

Computation relies on efficient C++ routines and includes multiple options for variable selection and sparse loading estimation.

S3 methods for objects of class `spca` include:
\tabular{ll}{
\code{\link{print.spca}} \tab Print nonzero loadings.\cr
\code{\link{plot.spca}} \tab Plot variance explained and sparse loadings.\cr
\code{\link{summary.spca}} \tab Summarize the solution with key statistics.\cr
}

Additional utilities include:
\tabular{ll}{
\code{\link{showload}} \tab prints, the nonzero loadings or contributions separatelyfor each sPC.\cr
\code{\link{compare.spca}} \tab Compare multiple `spca` objects (summaries and plots).\cr
\code{\link{aggregate_by_scale}} \tab Aggregate loadings/contributions by scale.\cr
}

The script to replicate the examples in the article is available at
\code{inst/scripts/Replicate_Examples.R} and can be located after installation via \code{system.file("scripts", "Replicate_Examples.R", package = "PMlsspca")}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/merolagio/PMlsspca}
  \item Report bugs at \url{https://github.com/merolagio/PMlsspca/issues}
}

}
\keyword{internal}
