% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods_For_Class.R
\name{summary.spca}
\alias{summary.spca}
\title{\strong{ Prints summaries from an spca object}}
\usage{
\method{summary}{spca}(
  spca_obj,
  cols,
  contribution = TRUE,
  variance_metrics = "cumulative_relative",
  minload = FALSE,
  rtn = FALSE,
  prn = TRUE,
  thrsehcard = 0.001
)
}
\arguments{
\item{spca_obj}{An spca object.}

\item{cols}{A vector indicating which components should be included. Default
all. If an integer is passed, it is set to 1:cols.}

\item{contribution}{Logical: should the loadings be standardised to unit L1 norm
(and printed as percentage contributions)}

\item{variance_metrics}{Character vector: which variance metrics to include.
Options: "relative" (RVEXP), "cumulative_relative" (RCVEXP), "both", or "none".
Default is "cumulative_relative".}

\item{minload}{Logical: should minimum loading/contribution be included?}

\item{rtn}{Logical: should the summary matrix be returned?}

\item{prn}{Logical: should anything be printed?}

\item{thrsehcard}{Value below which loadings are considered zero and not
counted in the cardinality}
}
\value{
If rtn = TRUE, a numerical matrix with the summaries.
}
\description{
Prints summaries and comparisons with the full PCA solutions for a set of LS
SPCA loadings.
}
\details{
The summaries are printed as formatted text, if rtn = TRUE, the value
returned is a numerical matrix.

For each component the following summaries are computed: \tabular{ll}{ 
VEXP\tab The percentage variance explained\cr
CVEXP \tab The percentage cumulative variance explained\cr
RVEXP \tab The variance explained relative to the corresponding PC\cr
RCVEXP \tab The cumulative variance explained relative to the corresponding PCs\cr
Card \tab The cardinality, that is the number of non zero loadings\cr
MinLoad/MinCont \tab Minimum absolute value of the non-zero loadings or contribution }
}
\seealso{
Examples in \code{\link{lsspca}}.
}
